fix: keep database code server-only
